Optimise uvars calculation
There is no magic to this.  The original code was doing uncoalesced
accesses to memory, and making them coalesced is simply much faster.

In the Ubiquitin case, the reductions themselves are tiny (about a
hundred or so elements each) and there's roughly 10,000 of them to be
done independently.  There's no real need for LDS because it isn't
necessary for threads to share any data.

For the Ubiqutin study, the original kernel saw approximately 60 GB/s
and this improved one sees approximately 1200 GB/s on an MI250X.  The
expectation is that the kernel ought to be bound by memory bandwidth.